+++ Editor/GraphEditor.cs
@@ -117,8 +117,41 @@ namespace Exodrifter.NodeGraph
 
 		#endregion
 
+		private Texture2D boxTexture;
+
 		void OnGUI()
 		{
+			// Make the box texture opaque
+			var oldTex = GUI.skin.box.normal.background;
+			if (boxTexture == null && EditorGUIUtility.isProSkin)
+			{
+				// Make a copy of the old texture
+				var tmp = RenderTexture.GetTemporary(oldTex.width, oldTex.height);
+
+				Graphics.Blit(oldTex, tmp);
+				RenderTexture previous = RenderTexture.active;
+				RenderTexture.active = tmp;
+				boxTexture = new Texture2D(oldTex.width, oldTex.height);
+				boxTexture.ReadPixels(new Rect(0, 0, tmp.width, tmp.height), 0, 0);
+				RenderTexture.active = previous;
+				RenderTexture.ReleaseTemporary(tmp);
+
+				// Remove alpha
+				var colors = boxTexture.GetPixels();
+				for (int i = 0; i < colors.Length; ++i)
+				{
+					// Pro background color is RGB(64, 64, 64)
+					colors[i].r = 0.2196f + (colors[i].r * colors[i].a);
+					colors[i].g = 0.2196f + (colors[i].g * colors[i].a);
+					colors[i].b = 0.2196f + (colors[i].b * colors[i].a);
+					colors[i].a = 1;
+				}
+
+				boxTexture.SetPixels(colors);
+				boxTexture.Apply();
+			}
+			GUI.skin.box.normal.background = boxTexture;
+
 			rectCache = rectCache ?? new Dictionary<Socket, Rect>();
 			rectCache.Clear();
 
@@ -262,6 +295,8 @@ namespace Exodrifter.NodeGraph
 				search.OnGUI(this);
 
 				GUI.EndClip();
+
+				GUI.skin.box.normal.background = oldTex;
 			}
 
 			switch (Event.current.type)